@@ -19,39 +19,38 @@
     "lint": "gts check"
   },
   "dependencies": {
-    "chalk": "^3.0.0",
+    "chalk": "^4.0.0",
     "cheerio": "^1.0.0-rc.2",
     "finalhandler": "^1.1.2",
-    "gaxios": "^2.0.1",
-    "jsonexport": "^2.4.1",
-    "meow": "^6.0.0",
+    "gaxios": "^3.0.0",
+    "jsonexport": "^3.0.0",
+    "meow": "^7.0.0",
     "p-queue": "^6.2.1",
     "serve-static": "^1.14.1",
     "server-destroy": "^1.0.1",
-    "update-notifier": "^4.0.0"
+    "update-notifier": "^5.0.0"
   },
   "devDependencies": {
     "@types/chai": "^4.2.7",
     "@types/cheerio": "^0.22.10",
     "@types/finalhandler": "^1.1.0",
     "@types/meow": "^5.0.0",
-    "@types/mocha": "^7.0.0",
+    "@types/mocha": "^8.0.0",
     "@types/node": "^12.7.12",
     "@types/serve-static": "^1.13.3",
     "@types/server-destroy": "^1.0.0",
-    "@types/sinon": "^7.5.0",
-    "@types/update-notifier": "^4.0.0",
-    "assert-rejects": "^1.0.0",
+    "@types/sinon": "^9.0.0",
+    "@types/update-notifier": "^5.0.0",
     "c8": "^7.0.0",
     "chai": "^4.2.0",
     "codecov": "^3.6.1",
     "execa": "^4.0.0",
-    "gts": "^1.1.0",
-    "mocha": "^7.0.0",
-    "nock": "^12.0.0",
+    "gts": "^3.0.0",
+    "mocha": "^8.0.0",
+    "nock": "^13.0.0",
     "semantic-release": "^17.0.0",
     "sinon": "^9.0.0",
-    "typescript": "^3.6.4"
+    "typescript": "^4.0.0"
   },
   "engines": {
     "node": ">=10"
