@@ -3,13 +3,16 @@
 import * as meow from 'meow';
 import * as updateNotifier from 'update-notifier';
 import chalk = require('chalk');
-import { LinkChecker, LinkState, LinkResult, CheckOptions } from './index';
-import { promisify } from 'util';
-import { Flags, getConfig } from './config';
+import {LinkChecker, LinkState, LinkResult, CheckOptions} from './index';
+import {promisify} from 'util';
+import {Flags, getConfig} from './config';
+
+// eslint-disable-next-line @typescript-eslint/no-var-requires
 const toCSV = promisify(require('jsonexport'));
 
+// eslint-disable-next-line @typescript-eslint/no-var-requires
 const pkg = require('../../package.json');
-updateNotifier({ pkg }).notify();
+updateNotifier({pkg}).notify();
 
 const cli = meow(
   `
@@ -40,6 +43,9 @@ const cli = meow(
       --silent
           Only output broken links
 
+      --timeout
+          Request timeout in ms.  Defaults to 0 (no timeout).
+
       --help
           Show this command.
 
@@ -52,12 +58,13 @@ const cli = meow(
 `,
   {
     flags: {
-      config: { type: 'string' },
-      concurrency: { type: 'string' },
-      recurse: { type: 'boolean', alias: 'r' },
-      skip: { type: 'string', alias: 's' },
-      format: { type: 'string', alias: 'f' },
-      silent: { type: 'boolean' },
+      config: {type: 'string'},
+      concurrency: {type: 'number'},
+      recurse: {type: 'boolean', alias: 'r'},
+      skip: {type: 'string', alias: 's'},
+      format: {type: 'string', alias: 'f'},
+      silent: {type: 'boolean'},
+      timeout: {type: 'number'},
     },
     booleanDefault: undefined,
   }
@@ -107,6 +114,7 @@ async function main() {
   const opts: CheckOptions = {
     path: cli.input[0],
     recurse: flags.recurse,
+    timeout: Number(flags.timeout),
     concurrency: Number(flags.concurrency),
   };
   if (flags.skip) {
@@ -137,7 +145,7 @@ async function main() {
         acc[parent].push(curr);
       }
       return acc;
-    }, {} as { [index: string]: LinkResult[] });
+    }, {} as {[index: string]: LinkResult[]});
 
     Object.keys(parents).forEach(parent => {
       const links = parents[parent];
@@ -178,6 +186,7 @@ async function main() {
         )} links in ${chalk.cyan(total.toString())} seconds.`
       )
     );
+    // eslint-disable-next-line no-process-exit
     process.exit(1);
   }
 
